+++ CustomSabers/Data/Whacker Data/WhackerConfig.cs
@@ -1,0 +1,8 @@
+﻿namespace CustomSabersLite.Data;
+
+internal struct WhackerConfig
+{
+    public bool hasTrail;
+    public bool keepFakeGlow;
+    public bool isLegacy;
+}
