+++ CustomSabers/Data/Whacker Data/WhackerDescriptor.cs
@@ -1,0 +1,9 @@
+﻿namespace CustomSabersLite.Data;
+
+internal struct WhackerDescriptor
+{
+    public string objectName;
+    public string author;
+    public string description;
+    public string coverImage;
+}
