+++ CustomSabers/Data/Whacker Data/WhackerObject.cs
@@ -1,0 +1,9 @@
+﻿namespace CustomSabersLite.Data;
+
+internal struct WhackerObject
+{
+    public string androidFileName;
+    public string pcFileName;
+    public WhackerDescriptor descriptor;
+    public WhackerConfig config;
+}
