+++ CustomSabers/Data/Whacker Data/WhackerTrail.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+
+namespace CustomSabersLite.Data;
+
+internal struct WhackerTrail
+{
+    public int trailId;
+    public CustomSaber.ColorType colorType;
+    public Color trailColor;
+    public Color multiplierColor;
+    public int length;
+    public float whiteStep;
+}
