+++ CustomSabers/Data/Whacker Data/WhackerTrailTransform.cs
@@ -1,0 +1,7 @@
+﻿namespace CustomSabersLite.Data;
+
+internal struct WhackerTrailTransform
+{
+    public int trailId;
+    public bool isTop;
+}
